# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/admin-users": {
      "get": {
        "operationId": "AdminUsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Lister les comptes administrateurs (rôle, activation, dernière connexion, sessions vivantes).",
        "tags": [
          "admin-users"
        ]
      },
      "post": {
        "operationId": "AdminUsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        },
        "summary": "Créer un compte administrateur. Le mot de passe initial est transmis hors plateforme (aucun envoi d’e-mail — D-011).",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/admin-users/{adminUserId}": {
      "patch": {
        "description": "Refuse de vous désactiver ou de vous retirer votre propre rôle SUPER_ADMIN, et de retirer le DERNIER SUPER_ADMIN actif (la plateforme deviendrait inadministrable). Un changement de rôle ou une désactivation RÉVOQUE les sessions du compte : sans cela, le jeton déjà émis garderait les anciens droits jusqu’à son expiration.",
        "operationId": "AdminUsersController_update",
        "parameters": [
          {
            "name": "adminUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        },
        "summary": "Modifier un compte : nom, rôle, activation. Tracé dans l’AuditLog.",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/admin-users/{adminUserId}/password": {
      "post": {
        "operationId": "AdminUsersController_resetPassword",
        "parameters": [
          {
            "name": "adminUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        },
        "summary": "Réinitialiser le mot de passe d’un administrateur (le SUPER_ADMIN le pose et le transmet hors plateforme). Révoque ses sessions.",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/admin-users/{adminUserId}/sessions": {
      "get": {
        "description": "Ce n’est pas un journal de connexion complet : les tentatives ÉCHOUÉES ne sont enregistrées nulle part en base (`failedAttemptsRecorded: false`). Les journaliser demanderait une table dédiée et une écriture dans le chemin d’authentification — point ouvert, pas une donnée inventée.",
        "operationId": "AdminUsersController_sessions",
        "parameters": [
          {
            "name": "adminUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionsDto"
                }
              }
            }
          }
        },
        "summary": "Journal des SESSIONS d’un administrateur (ouverture, IP, navigateur, révocation), reconstitué depuis les jetons de rafraîchissement.",
        "tags": [
          "admin-users"
        ]
      }
    }
  },
  "info": {
    "title": "Najah Partners API",
    "description": "API backend — plateforme MLM (BV, e-cards, commissions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "enum": [
              "SUPER_ADMIN",
              "MANAGER",
              "SUPPORT"
            ],
            "type": "string",
            "description": "Rôle FIXE (enum). Les droits associés sont codés dans les guards du backend."
          },
          "active": {
            "type": "boolean",
            "description": "Un compte désactivé ne peut plus se connecter, et ses sessions en cours sont révoquées."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Dernière ouverture de session connue, déduite des jetons de rafraîchissement (voir le journal des sessions). `null` = ce compte ne s’est jamais connecté."
          },
          "activeSessionCount": {
            "type": "number",
            "description": "Sessions actuellement valides (jetons non révoqués, non expirés)."
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "active",
          "createdAt",
          "activeSessionCount"
        ]
      },
      "CreateAdminUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Sarra Ben Amor"
          },
          "email": {
            "type": "string",
            "maxLength": 180,
            "format": "email",
            "example": "sarra@najah-partners.tn"
          },
          "role": {
            "enum": [
              "SUPER_ADMIN",
              "MANAGER",
              "SUPPORT"
            ],
            "type": "string",
            "description": "Rôle FIXE. Les droits qui s’y attachent sont codés dans les guards du backend."
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "description": "Mot de passe initial, transmis hors plateforme (aucun envoi d’e-mail n’existe — D-011)."
          }
        },
        "required": [
          "name",
          "email",
          "role",
          "password"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Sarra Ben Amor"
          },
          "role": {
            "enum": [
              "SUPER_ADMIN",
              "MANAGER",
              "SUPPORT"
            ],
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "description": "Désactiver révoque aussi les sessions en cours : sans cela, le compte resterait utilisable jusqu’à l’expiration de son jeton."
          }
        }
      },
      "ResetAdminPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "description": "Nouveau mot de passe. Les sessions en cours du compte sont révoquées."
          }
        },
        "required": [
          "password"
        ]
      },
      "AdminSessionDto": {
        "type": "object",
        "properties": {
          "familyId": {
            "type": "string",
            "description": "Identifiant de la famille de jetons = une session."
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Ouverture de la session (première émission de la famille)."
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "description": "Dernière rotation du jeton — l’activité la plus récente."
          },
          "ip": {
            "type": "string",
            "nullable": true,
            "description": "Adresse IP vue par le serveur."
          },
          "userAgent": {
            "type": "string",
            "nullable": true,
            "description": "Navigateur déclaré (User-Agent)."
          },
          "current": {
            "type": "boolean",
            "description": "Session encore utilisable : au moins un jeton non révoqué et non expiré. Faux = déconnexion, rotation terminée, ou expiration."
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "familyId",
          "startedAt",
          "lastSeenAt",
          "current",
          "expiresAt"
        ]
      },
      "AdminSessionsDto": {
        "type": "object",
        "properties": {
          "adminUserId": {
            "type": "number"
          },
          "sessions": {
            "description": "Plus récentes d’abord.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSessionDto"
            }
          },
          "failedAttemptsRecorded": {
            "type": "boolean",
            "description": "Les tentatives de connexion ÉCHOUÉES ne sont enregistrées nulle part : ce journal ne les contient pas (point ouvert)."
          }
        },
        "required": [
          "adminUserId",
          "sessions",
          "failedAttemptsRecorded"
        ]
      }
    }
  }
}